Remove MigrationTaskService and NodeGroupMigrationTaskService
migration task service grew to be incredibly complex. Instead of using a generic task, that is parametric on so many things, developers can add migration code in the service itself: v2 of a service handleStart, handlePut, handlePatch simply needs to do an on demand change to the new document format.

as an alternative, a new service can be written, that specifically migrates ServiceX into ServiceY: the migration service can run on a remote node, outside the live node group and perform whatever operations it needs, while the cluster is live